Working queuing on redis with MPI
