+++ .gitignore
@@ -105,7 +105,7 @@ ENV/
 \.vscode/
 
 # users data
-static/DATA/
+Project/Server/DATA/
 
 # db files
 *.rdb
